rag-project/requirements.txt
chromadb==1.5.2
groq==1.0.0
pypdf==6.7.4
python-dotenv==1.2.1
sentence_transformers==5.2.3
streamlit==1.54.0
